consolidate live-view into target_detect.py
